requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
apscheduler>=3.10.4
slack-sdk>=3.27.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.127.0
openai>=1.30.0
httpx>=0.27.0
python-dotenv>=1.0.0
pytz>=2024.1
python-multipart>=0.0.9
aiohttp>=3.9.0
